@@ -27,9 +27,8 @@ require __DIR__.'/../vendor/autoload.php';
 |
 */
 
-$compiledPath = __DIR__.'/../storage/framework/compiled.php';
+$compiledPath = __DIR__.'/cache/compiled.php';
 
-if (file_exists($compiledPath))
-{
-	require $compiledPath;
+if (file_exists($compiledPath)) {
+    require $compiledPath;
 }
