@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Laravel - A PHP Framework For Web Artisans
  *
@@ -39,17 +40,17 @@ $app = require_once __DIR__.'/../bootstrap/app.php';
 | Run The Application
 |--------------------------------------------------------------------------
 |
-| Once we have the application, we can simply call the run method,
-| which will execute the request and send the response back to
+| Once we have the application, we can handle the incoming request
+| through the kernel, and send the associated response back to
 | the client's browser allowing them to enjoy the creative
 | and wonderful application we have prepared for them.
 |
 */
 
-$kernel = $app->make('Illuminate\Contracts\Http\Kernel');
+$kernel = $app->make(Illuminate\Contracts\Http\Kernel::class);
 
 $response = $kernel->handle(
-	$request = Illuminate\Http\Request::capture()
+    $request = Illuminate\Http\Request::capture()
 );
 
 $response->send();
