@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Laravel - A PHP Framework For Web Artisans
  *
@@ -7,15 +8,14 @@
  */
 
 $uri = urldecode(
-	parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH)
+    parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH)
 );
 
 // This file allows us to emulate Apache's "mod_rewrite" functionality from the
 // built-in PHP web server. This provides a convenient way to test a Laravel
 // application without having installed a "real" web server software here.
-if ($uri !== '/' and file_exists(__DIR__.'/public'.$uri))
-{
-	return false;
+if ($uri !== '/' && file_exists(__DIR__.'/public'.$uri)) {
+    return false;
 }
 
 require_once __DIR__.'/public/index.php';
